requirements.txt
# Core ML packages
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
tensorflow>=2.18.0
tf-keras
sentence-transformers>=2.2.2
transformers>=4.36.0

# Graph & Retrieval
neo4j>=5.12.0
faiss-cpu>=1.7.4
networkx>=3.1
node2vec>=0.4.3

# NLP
spacy>=3.7.0
pypdf>=3.15.0

# Google Gemini
google-genai>=0.6.1
